+++ src/aim/controllers/ctl_s3.py
@@ -1,6 +1,8 @@
 import os
 from aim.core.exception import StackException
 from aim.core.exception import AimErrorCode
+from aim.models import schemas
+from aim.models.locations import get_parent_by_interface
 from aim.stack_group import StackGroup
 from aim.controllers.controllers import Controller
 from botocore.exceptions import ClientError
@@ -43,8 +45,6 @@ class S3Context():
             'group_id': None,
             'config': None,
             'ref': resource_ref,
-            'bucket_name_prefix': None,
-            'bucket_name_suffix': None,
             'stack': None
         }
 
@@ -53,67 +53,76 @@ class S3Context():
             return
         self.bucket_context['stack'].add_hooks(stack_hooks)
 
-    def add_stack(  self,
-                    bucket_policy_only=False,
-                    stack_hooks=None,
-                    new_stack=True,
-                    stack_tags=None):
-        s3_template = aim.cftemplates.S3(self.aim_ctx,
-                                         self.account_ctx,
-                                         self.region,
-                                         self.stack_group,
-                                         stack_tags,
-                                         stack_hooks,
-                                         self.bucket_context,
-                                         bucket_policy_only,
-                                         self.resource_ref)
-
+    def add_stack(
+        self,
+        bucket_policy_only=False,
+        stack_hooks=None,
+        new_stack=True,
+        stack_tags=None
+    ):
+        s3_template = aim.cftemplates.S3(
+            self.aim_ctx,
+            self.account_ctx,
+            self.region,
+            self.stack_group,
+            stack_tags,
+            stack_hooks,
+            self.bucket_context,
+            bucket_policy_only,
+            self.resource_ref
+        )
         if bucket_policy_only == False:
             if self.bucket_context['stack'] != None:
                 raise StackException(AimErrorCode.Unknown)
             self.bucket_context['stack'] = s3_template.stack
 
-    def add_bucket( self,
-                    region,
-                    bucket_id,
-                    bucket_group_id,
-                    bucket_name_prefix,
-                    bucket_name_suffix,
-                    bucket_config,
-                    stack_hooks=None):
-
+    def add_bucket(
+        self,
+        bucket,
+        bucket_name_prefix=None,
+        bucket_name_suffix=None,
+        stack_hooks=None
+    ):
+        "Add a bucket: will create a stack and stack hooks as needed"
         if self.bucket_context['id'] != None:
             print("Bucket already exists: %s" % (self.resource_ref))
             raise StackException(AimErrorCode.Unknown)
 
-        if bucket_name_suffix == None:
-            bucket_name_suffix = vocabulary.aws_regions[region]['short_name']
-        else:
-            bucket_name_suffix += '-'+vocabulary.aws_regions[region]['short_name']
+        if bucket_name_prefix:
+            bucket.bucket_name_prefix = bucket_name_prefix
+        if bucket_name_suffix:
+            bucket.bucket_name_suffix = bucket_name_suffix
 
-        self.bucket_context['id'] = bucket_id
-        self.bucket_context['group_id'] = bucket_group_id
-        self.bucket_context['config'] = bucket_config
+        self.bucket_context['id'] = bucket.name
+        res_group = get_parent_by_interface(bucket, schemas.IResourceGroup)
+        if res_group != None:
+            self.bucket_context['group_id'] = res_group.name
+        self.bucket_context['config'] = bucket
         self.bucket_context['ref'] = self.resource_ref
-        self.bucket_context['bucket_name_prefix'] = bucket_name_prefix
-        self.bucket_context['bucket_name_suffix'] = bucket_name_suffix
+        bucket.resolve_ref_obj = self
 
-        bucket_config.resolve_ref_obj = self
 
-        # If the bucket already exists, do not create a stack for it
-        if self.bucket_context['config'].external_resource == False:
-            # S3 Delete on Stack Delete hook
-            utils.log_action_col("Init", "S3", "Bucket", bucket_id+": "+self.get_bucket_name(), False, bucket_config.is_enabled())
+        if bucket.external_resource == True:
+            # if the bucket already exists, do not create a stack for it
+            utils.log_action_col(
+                "Init", "S3", "External Bucket", bucket.name + ": " + self.get_bucket_name(), False, bucket.is_enabled()
+            )
+        else:
+            utils.log_action_col(
+                "Init", "S3", "Bucket", bucket.name + ": " + bucket.get_bucket_name(), False, bucket.is_enabled()
+            )
             if stack_hooks == None:
                 stack_hooks = StackHooks(self.aim_ctx)
-            stack_hooks.add('S3StackGroup', 'delete', 'post',
-                            self.stack_hook_post_delete, None, self.bucket_context)
-            self.add_stack(bucket_policy_only=False,
-                            stack_hooks=stack_hooks,
-                            stack_tags=self.stack_tags)
-        else:
-            utils.log_action_col("Init", "S3", "External Bucket", bucket_id+": "+self.get_bucket_name(), False, bucket_config.is_enabled())
-
+            # S3 Delete on Stack Delete hook
+            stack_hooks.add(
+                'S3StackGroup', 'delete', 'post',
+                self.stack_hook_post_delete, None, self.bucket_context
+            )
+            self.add_stack(
+                bucket_policy_only=False,
+                stack_hooks=stack_hooks,
+                stack_tags=self.stack_tags
+            )
 
     def add_bucket_policy(self, policy_dict, stack_hooks=None, new_stack=True):
         bucket_config = self.bucket_context['config']
@@ -123,26 +132,17 @@ class S3Context():
             for policy in bucket_config.policy:
                 policy.processed = True
         bucket_config.add_policy(policy_dict)
-
-        self.add_stack( bucket_policy_only=True,
-                        stack_hooks=stack_hooks,
-                        stack_tags=self.stack_tags)
-
-    def get_bucket_name(self):
-        if self.bucket_context['config'].external_resource == True:
-            return self.bucket_context['config'].bucket_name
-        else:
-            bucket_name = '-'.join([self.bucket_context['bucket_name_prefix'],
-                                    self.bucket_context['config'].name,
-                                    self.bucket_context['config'].bucket_name,
-                                    self.bucket_context['bucket_name_suffix']])
-            return bucket_name.replace('_', '-').lower()
+        self.add_stack(
+            bucket_policy_only=True,
+            stack_hooks=stack_hooks,
+            stack_tags=self.stack_tags
+        )
 
     def get_bucket_arn(self):
-        return 'arn:aws:s3:::'+self.get_bucket_name()
+        return 'arn:aws:s3:::' + self.bucket_context['config'].get_bucket_name()
 
     def get_bucket_url(self):
-        return self.get_bucket_name()+'.s3.amazonaws.com'
+        return self.bucket_context['config'].get_bucket_name() + '.s3.amazonaws.com'
 
     def get_region(self):
         return self.region
@@ -160,12 +160,12 @@ class S3Context():
             self.stack_group.delete()
 
     def stack_hook_post_delete(self, hook, hook_arg):
-        # Empty the S3 Bucket if enabled
+        "Empty the S3 Bucket if enabled"
         bucket_context = hook_arg
         s3_config = bucket_context['config']
         s3_resource = self.account_ctx.get_aws_resource('s3', self.region)
         deletion_policy = s3_config.deletion_policy
-        bucket_name = self.get_bucket_name()
+        bucket_name = s3_config.get_bucket_name()
         if deletion_policy == "delete":
             print("Deleting S3 Bucket: %s" % (bucket_name))
             bucket = s3_resource.Bucket(bucket_name)
@@ -183,10 +183,8 @@ class S3Context():
         if self.bucket_context == None:
             print("ctl_s3: empty_bucket: ERROR: Unable to locate stack group for group: " + group_id)
             raise StackException(AimErrorCode.Unknown)
-
-
         s3_client = self.account_ctx.get_aws_client('s3')
-        bucket_name = self.get_bucket_name()
+        bucket_name = self.bucket_context['config'].get_bucket_name()
         try:
             response = s3_client.list_objects_v2(Bucket=bucket_name)
         except ClientError as e:
@@ -210,7 +208,7 @@ class S3Context():
         if ref.last_part == 'arn':
             return self.get_bucket_arn()
         elif ref.last_part == 'name':
-            return self.get_bucket_name()
+            return self.bucket_context['config'].get_bucket_name()
         elif ref.last_part == 'url':
             return self.get_bucket_url()
         elif ref.last_part == 'origin_id':
@@ -220,11 +218,11 @@ class S3Context():
 
 class S3Controller(Controller):
     def __init__(self, aim_ctx):
-        super().__init__(aim_ctx,
-                         "S3",
-                         "Resource")
-
-        #self.aim_ctx.log("S3 Service: Configuration: %s" % (name))
+        super().__init__(
+            aim_ctx,
+            "S3",
+            "Resource"
+        )
         self.contexts = {}
         self.init_s3_resource_done = False
 
@@ -233,27 +231,23 @@ class S3Controller(Controller):
             # Each bucket gets its own stack
             for bucket_id, bucket_config in env_config['buckets']:
                 resource_ref = 'resource.s3.buckets.{0}.{1}.{2}'.format(env_config['account_ctx'].get_name(), env_config['region'], bucket_id)
-                env_stack_group = S3StackGroup( self.aim_ctx,
-                                                env_config['account_ctx'],
-                                                env_config['region'],
-                                                'bucket',
-                                                self,
-                                                resource_ref,
-                                                stack_hooks=None)
-                self.init_context(  env_config['account_ctx'],
-                                    env_config['region'],
-                                    resource_ref,
-                                    env_stack_group,
-                                    stack_tags )
-                self.add_bucket(resource_ref,
-                                env_config['region'],
-                                bucket_id,
-                                None,
-                                'aim-s3',
-                                None,
-                                bucket_config,
-                                None)
-
+                env_stack_group = S3StackGroup(
+                    self.aim_ctx,
+                    env_config['account_ctx'],
+                    env_config['region'],
+                    'bucket',
+                    self,
+                    resource_ref,
+                    stack_hooks=None
+                )
+                self.init_context(
+                    env_config['account_ctx'],
+                    env_config['region'],
+                    resource_ref,
+                    env_stack_group,
+                    stack_tags
+                )
+                self.add_bucket(bucket_config)
 
     def init_s3_resource(self, controller_args, stack_tags):
         if self.init_s3_resource_done == True:
@@ -261,8 +255,8 @@ class S3Controller(Controller):
         utils.log_action_col("Init", "S3")
         self.init_s3_resource_done = True
         s3_env_map = {}
-        for bucket_id in self.aim_ctx.project['s3'].buckets.keys():
-            bucket_config = self.aim_ctx.project['s3'].buckets[bucket_id]
+        for bucket_id in self.aim_ctx.project['resource']['s3'].buckets.keys():
+            bucket_config = self.aim_ctx.project['resource']['s3'].buckets[bucket_id]
             account_ctx = self.aim_ctx.get_account_context(account_ref=bucket_config.account)
             region = bucket_config.region
             s3_env_id = '-'.join([account_ctx.get_name(), region])
@@ -292,8 +286,11 @@ class S3Controller(Controller):
             # and still do a lookup on them
             self.contexts['aim.ref '+resource_ref] = self.contexts[resource_ref]
 
-    def add_bucket(self, resource_ref, *args, **kwargs):
-        return self.contexts[resource_ref].add_bucket(*args, **kwargs)
+    def add_bucket(self, bucket, config_ref=None, **kwargs):
+        if config_ref:
+            return self.contexts[config_ref].add_bucket(bucket, **kwargs)
+        else:
+            return self.contexts[bucket.aim_ref].add_bucket(bucket, **kwargs)
 
     def add_bucket_policy(self, resource_ref, *args, **kwargs):
         return self.contexts[resource_ref].add_bucket_policy(*args, **kwargs)
@@ -305,7 +302,7 @@ class S3Controller(Controller):
         return self.contexts[resource_ref].get_bucket_arn(*args, **kwargs)
 
     def get_bucket_name(self, resource_ref, *args, **kwargs):
-        return self.contexts[resource_ref].get_bucket_name(*args, **kwargs)
+        return self.contexts[resource_ref].bucket_context['config'].get_bucket_name()
 
     def empty_bucket(self, resource_ref, *args, **kwargs):
         return self.contexts[resource_ref].empty_bucket(*args, **kwargs)